added SOME code to save video
code to save video is taken from a stackoverflow page linked in videorecord.c.
this code is 100% functional on its own. after adding this code to the graphics_module.cpp,
i started getting weird errors involving bits of code not related to the code i added.
i did a lot of debugging but it seems to be an external issue.
-adri