worked on saving qr code as image.
so it turns out that's really hard though. the url is given from the generateQR.js code in /src,
but can't directly be passed into the c++ code since we're just executing generateQR.js from system calls.
it's relatively easy to generate QR codes in javascript (since there are many many libraries),
but we have no way of saving those images natively or viewing them, since .js code is meant to be run
in a browser. that is, all the methods of viewing the QR codes it generates don't work for saving to storage,
since there is no version of local storage when executing .js code in a browser.

to try to get around this i just output the url to the file in a garbage text file, and hoped that it
would be easier to read the text file and generate the QR code from somewhere in the C++.
it was not easier. there are lots of attempts to make it easy to view this, but none *really* natively supported
by c++, and also none that could be implemented without really goofing up the structure of our program
anyways. will continue to work on it later.
-adri